+++ sdk/src/main/java/com/herewhite/sdk/internal/PostMessageCallback.java
@@ -1,0 +1,7 @@
+package com.herewhite.sdk.internal;
+
+import org.json.JSONObject;
+
+public interface PostMessageCallback {
+    void onMessage(JSONObject jsonObject);
+}